Two things are certain about science.  It does not stand still for long, and it is never boring.  Oh, among some poor souls, including even intellectuals in fields of high scholarship, science is frequently misperceived.  Many see it as only a body of facts, promulgated from on high in must, unintelligible textbooks, a collection of unchanging precepts defended with authoritarian vigor.  Others view it as nothing but a cold, dry narrow, plodding, rule-bound process -- the scientific method: hidebound, linear, and left brained.
These people are the victims of their own stereotypes.  They are
destined to view the world of science with a set of blinders.  They
know nothing of the tumult, cacophony, rambunctiousness, and
tendentiousness of the actual scientific process, let alone the
creativity, passion, and joy of discovery.  And they are likely to
know little of the continual procession of new insights and discoveries
that every day, in some way, change our view (if not theirs) of the
natural world.

		-- Kendrick Frazier, "The Year in Science: An Overview," in
        1988 Yearbook of Science and the Future, Encyclopaedia Britannica, Inc.